bot/requirements.txt
discord.py>=2.3.2
openai>=1.30.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiohttp>=3.9.0